+++ SMS/src/app/student/dashboard/dashboard.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { DialogService } from 'src/app/shared/dialog.service';
 
 @Component({
   selector: 'app-dashboard',
@@ -7,15 +8,23 @@ import { Router } from '@angular/router';
   styleUrls: ['./dashboard.component.css']
 })
 export class DashboardComponent implements OnInit {
-
+  userName:any
   constructor(
-    private route:Router
+    private route:Router,
+    private dialogService: DialogService
     ) { }
 
   ngOnInit(): void {
+    this.userName = localStorage.getItem('userName')
   }
-  logout(){
-    localStorage.clear();
-    this.route.navigateByUrl("/Login");
 
-}}
+  logout() {
+    this.dialogService.openConfirmDialog()
+      .afterClosed().subscribe(res => {
+        if (res) {
+          localStorage.clear();
+          this.route.navigateByUrl('/Login')
+        }
+      })
+  }  
+}
