+++ SMS/src/app/teacher/teacher-dashboard/teacher-dashboard.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { DialogService } from 'src/app/shared/dialog.service';
 
 @Component({
   selector: 'app-teacher-dashboard',
@@ -7,16 +8,22 @@ import { Router } from '@angular/router';
   styleUrls: ['./teacher-dashboard.component.css']
 })
 export class TeacherDashboardComponent implements OnInit {
-
-
+userName:any
   constructor(
-    private route:Router
+    private route:Router,
+    private dialogService: DialogService
     ) { }
   ngOnInit(): void {
-  }
-  logout(){
-    localStorage.clear();
-    this.route.navigateByUrl("/Login");
+    this.userName = localStorage.getItem('userName')
   }
 
+  logout() {
+    this.dialogService.openConfirmDialog()
+      .afterClosed().subscribe(res => {
+        if (res) {
+          localStorage.clear();
+          this.route.navigateByUrl('/Login')
+        }
+      })
+  }
 }
